+++ packages/@dicebear/big-ears-neutral/src/components/index.ts
@@ -1,0 +1,4 @@
+export { mouth } from './mouth';
+export { eyes } from './eyes';
+export { cheek } from './cheek';
+export { nose } from './nose';