+++ packages/@dicebear/big-ears/src/components/cheek.ts
@@ -1,0 +1,52 @@
+import type { ComponentGroup, ComponentPickCollection, ColorPickCollection } from '../static-types';
+
+export const cheek: ComponentGroup = {
+  'variant06': (components: ComponentPickCollection, colors: ColorPickCollection) => `
+  <g fill="#F88888">
+    <rect x="21" y="16" width="43.401" height="43.401" rx="21.701"/>
+    <rect x="194" y="16" width="43.401" height="43.401" rx="21.701"/>
+  </g>
+`,
+  'variant05': (components: ComponentPickCollection, colors: ColorPickCollection) => `
+  <g fill="#FFC6C6">
+    <rect x="21" y="16" width="43.401" height="43.401" rx="21.701"/>
+    <rect x="194" y="16" width="43.401" height="43.401" rx="21.701"/>
+  </g>
+`,
+  'variant04': (components: ComponentPickCollection, colors: ColorPickCollection) => `
+  <g fill="#F88888">
+    <rect x="11" y="16" width="56.261" height="43.401" rx="21.701"/>
+    <rect x="191" y="16" width="56.261" height="43.401" rx="21.701"/>
+  </g>
+`,
+  'variant03': (components: ComponentPickCollection, colors: ColorPickCollection) => `
+  <g fill="#FFC6C6">
+    <rect x="11" y="16" width="56.261" height="43.401" rx="21.701"/>
+    <rect x="191" y="16" width="56.261" height="43.401" rx="21.701"/>
+  </g>
+`,
+  'variant02': (components: ComponentPickCollection, colors: ColorPickCollection) => `
+  <g fill="#000" fill-opacity=".4">
+    <rect x="33" y="38.467" width="12.86" height="12.86" rx="6.43"/>
+    <rect x="57.112" y="24" width="12.86" height="14.467" rx="6.43"/>
+    <rect x="76.401" y="38.467" width="12.86" height="12.86" rx="6.43"/>
+    <g>
+      <rect x="166.419" y="38.467" width="12.86" height="12.86" rx="6.43"/>
+      <rect x="190.531" y="24" width="12.86" height="14.467" rx="6.43"/>
+      <rect x="209.821" y="38.467" width="12.86" height="12.86" rx="6.43"/>
+    </g>
+  </g>
+`,
+  'variant01': (components: ComponentPickCollection, colors: ColorPickCollection) => `
+  <g fill="#000" fill-opacity=".19">
+    <rect x="33" y="38.467" width="12.86" height="12.86" rx="6.43"/>
+    <rect x="57.112" y="24" width="12.86" height="14.467" rx="6.43"/>
+    <rect x="76.401" y="38.467" width="12.86" height="12.86" rx="6.43"/>
+    <g>
+      <rect x="166.419" y="38.467" width="12.86" height="12.86" rx="6.43"/>
+      <rect x="190.531" y="24" width="12.86" height="14.467" rx="6.43"/>
+      <rect x="209.821" y="38.467" width="12.86" height="12.86" rx="6.43"/>
+    </g>
+  </g>
+`,
+}